+++ WHATSNEW.md
@@ -1,5 +1,7 @@
 ### Added
 
+- Method inventory.get_item() to only get one single item from the inventory
+
 ### Changed
 
 ### Removed